refactor: keep IdValData local to vec_progress tests
# Summary

Move the test-only `IdValData` fixture out of the shared progress helper and define it beside the vec_progress tests that need application-owned data.

# Details

The shared progress entry helper now contains only `IdVal`, reducing test-only surface in the progress module.

The local fixture keeps `VecProgressEntryData` coverage unchanged while dropping the unused `Display` implementation.
